Update v0.2.0
Updated command handler and new command file template. The new template is easier to read and work with.

Added new 'get-commands' command. Will send all the registered slash commands with their id. This command itself is not a slash command.

Added 'config.json' for easier distribution later. The token is still in the hidden '.env' file, however the prefix and development guild are new editable in the config file.

Added the ability to respond to direct messages. Slash commands are not yet useable in dm's.